add syscall names, make stuff threadsafe
previously an int var in the debugger state struct decided upon
whether we return from or enter a syscall.
with new threads spawned, this couldn't work.

i generated tables containing the syscall names and argument count
from strace headers.
unlike in strace, they are all in .rodata and efficiently packed.

with these changes, the demo program syscallhook is successfully
able to hook any mmap() syscalls in such a way that the flag
MAP_32BIT is added to all of them. this makes it possible
to use

https://github.com/rofl0r/ags/commit/21ff9983b815f9185cb5a1cbb4d728c2ec0a7da2

to play AGS on linux AMD64 without any crashes.