revamp API to use uintptr_t for pointers to tracee memory
in C, a pointer needs to point to some actual memory in the process'
address space. since these are addresses in another process' address
space, using them is invalid, and additionally, required casts
all over the place. designing the API with uintptr_t makes it a lot
neater, plus one can easily see from a function signature whether
we deal with a local pointer or a foreign one.